Work towards #1255.

(Still needs debugging, @oskooi.)