@oskooi 2020-06-24
After applying the bug fix to `src/mpb.cpp:71` described above and verifying that all tests in the `make check` suite pass, a benchmarking test for this PR involving a large 3d simulation with a 2d source plane (ridge waveguide cross section) reveals that there is practically *no* speed up relative to master. The test is performed on a single machine (i.e., no networking/MPI) with 14 processors/chunks.

The test for master involves timing the call to `set_maxwell_dielectric` in `src/mpb.cpp:385` via the `wall_time()` function:

https://github.com/NanoComp/meep/blob/80edb8170a6fd99eb6f26c7f72425b1ee917222e/src/mpb.cpp#L385

**master**
```
set_maxwell_dielectric:, 1067.82 s
```

The test for this PR involves timing each of the two calls to `set_maxwell_dielectric` as well as `sum_to_all` which is called in between the two calls to `set_maxwell_dielectric`.

**this PR**
```
set_maxwell_dielectric1:, 0.661123 s
sum_to_all:, 1067.43 s
set_maxwell_dielectric2:, 0.000506878 s
```
These results demonstrate that the `sum_to_all` call in this PR is taking just as long as the single call to `set_maxwell_dielectric` in master (even though `set_maxwell_dielectric` has been sped up considerably).