@stevengj 2020-06-24
You might try just timing an `all_wait();` call right before the `sum_to_all()`, to see if the wait time is just due to one process taking a long time to reach that point.