@oskooi 2020-06-24
Putting an `all_wait();` right before `sum_to_all()` and timing each function call separately (with the output displayed using `master_printf`) reveals that it is the `all_wait()` that is taking up most of the time:

```
set_maxwell_dielectric1:, 0.664365 s
all_wait:, 1059.81 s
sum_to_all:, 0.004601 s
set_maxwell_dielectric2:, 0.000252962 s
```
To investigate whether one or more of the chunks is causing the delay, each of the 14 chunks (rank 0-13) outputs its wall time for `all_time()` separately via `printf`:
```
all_wait:, 0 (rank), 1060.05 s
all_wait:, 1 (rank), 564.322 s
all_wait:, 2 (rank), 533.729 s
all_wait:, 3 (rank), 495.724 s
all_wait:, 4 (rank), 942.212 s
all_wait:, 5 (rank), 29.6588 s
all_wait:, 6 (rank), 735.957 s
all_wait:, 7 (rank), 552.688 s
all_wait:, 8 (rank), 493.958 s
all_wait:, 9 (rank), 1026.89 s
all_wait:, 10 (rank), 0.000249147 s
all_wait:, 11 (rank), 710.75 s
all_wait:, 12 (rank), 589.994 s
all_wait:, 13 (rank), 953.669 s
```
These results indicate that while there is one chunk (rank 0) causing the delay there are several other chunks with comparable times (i.e., 4, 9, 13).