@stevengj 2020-06-26
Try putting an `all_wait()` at the beginning of `get_eigenmode` as well, to check whether the synchronization delay originates in this function or somewhere else.

Meanwhile, I'm going to merge this anyway, since it should scale better to do things this way, and tests pass.